requirements/local.txt
asgiref==3.5.2
Django==3.1.4
djangorestframework==3.13.1
pkg_resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
Unipath==1.1
